Refactor verifyVSCodeConfig into separate files
And move ProjectSettings.ts into the same folder
