+++ src/vsCodeConfig/ITask.ts
@@ -1,0 +1,18 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { TaskDefinition } from "vscode";
+
+export interface ITask extends TaskDefinition {
+    label?: string;
+    options?: ITaskOptions;
+}
+
+export interface ITaskOptions {
+    cwd?: string;
+    env?: {
+        [key: string]: string;
+    };
+}
