+++ test/runWithSetting.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { extensionPrefix, getGlobalFuncExtensionSetting, updateGlobalSetting } from "../extension.bundle";
+import { extensionPrefix, getGlobalSetting, updateGlobalSetting } from "../extension.bundle";
 
 export async function runWithFuncSetting(key: string, value: string | undefined, callback: () => Promise<void>): Promise<void> {
     await runWithSettingInternal(key, value, extensionPrefix, callback);
@@ -14,7 +14,7 @@ export async function runWithSetting(key: string, value: string | undefined, cal
 }
 
 async function runWithSettingInternal(key: string, value: string | undefined, prefix: string, callback: () => Promise<void>): Promise<void> {
-    const oldValue: string | undefined = getGlobalFuncExtensionSetting(key, prefix);
+    const oldValue: string | undefined = getGlobalSetting(key, prefix);
     try {
         await updateGlobalSetting(key, value, prefix);
         await callback();
